+++ tests/testthat/test-iter.R
@@ -45,10 +45,7 @@ test_that("NULL terminates iterators", {
   done <- TRUE
   expect_null(it())
   expect_true(is_done(it))
-  expect_error(
-    it(),
-    class = "flowery_error_iterator_done"
-  )
+  expect_null(it())
 })
 
 test_that("boxed NULLs don't terminate iterators", {